feat: added relations between posts and users, some fixes in routing

+++ posts/repository.go
@@ -9,9 +9,9 @@ import (
 type Repository interface {
 	GetAll() ([]Post, error)
 	GetById(postId int) (*Post, error)
-	CreatePost(newPost Post) (string, error)
-	UpdatePost(newPost Post, postId int) (string, error)
-	DeletePost(postId int) (string, error)
+	CreatePost(newPost Post, userId int) (string, error)
+	UpdatePost(newPost Post, postId int, userId int) (string, error)
+	DeletePost(postId int, userId int) (string, error)
 }
 
 type repositoryImpl struct {
@@ -25,7 +25,7 @@ func NewRepository(db *sql.DB) Repository {
 func (r *repositoryImpl) GetAll() ([]Post, error) {
 	posts := []Post{}
 
-	rows, err := r.db.Query("SELECT id, title, content FROM posts")
+	rows, err := r.db.Query("SELECT id, title, content, user_id FROM posts")
 	if err != nil {
 		return nil, err
 	}
@@ -33,7 +33,7 @@ func (r *repositoryImpl) GetAll() ([]Post, error) {
 
 	for rows.Next() {
 		var post Post
-		err := rows.Scan(&post.ID, &post.Title, &post.Content)
+		err := rows.Scan(&post.ID, &post.Title, &post.Content, &post.UserId)
 		if err != nil {
 			return nil, err
 		}
@@ -46,7 +46,7 @@ func (r *repositoryImpl) GetAll() ([]Post, error) {
 func (r *repositoryImpl) GetById(postId int) (*Post, error) {
 	var post Post
 
-	err := r.db.QueryRow("SELECT id, title, content FROM posts WHERE id = ?", postId).Scan(&post.ID, &post.Title, &post.Content)
+	err := r.db.QueryRow("SELECT id, title, content, user_id FROM posts WHERE id = ?", postId).Scan(&post.ID, &post.Title, &post.Content, &post.UserId)
 	if err == sql.ErrNoRows {
 		return nil, errors.New("Post not found")
 	} else if err != nil {
@@ -56,8 +56,8 @@ func (r *repositoryImpl) GetById(postId int) (*Post, error) {
 	return &post, nil
 }
 
-func (r *repositoryImpl) CreatePost(newPost Post) (string, error) {
-	result, err := r.db.Exec("INSERT INTO posts (title, content) VALUES (?, ?)", newPost.Title, newPost.Content)
+func (r *repositoryImpl) CreatePost(newPost Post, userId int) (string, error) {
+	result, err := r.db.Exec("INSERT INTO posts (title, content, user_id) VALUES (?, ?, ?)", newPost.Title, newPost.Content, userId)
 	if err != nil {
 		return "", err
 	}
@@ -66,8 +66,8 @@ func (r *repositoryImpl) CreatePost(newPost Post) (string, error) {
 	return fmt.Sprintf("Post created successfully with ID %d", id), nil
 }
 
-func (r *repositoryImpl) UpdatePost(updatedPost Post, postId int) (string, error) {
-	result, err := r.db.Exec("UPDATE posts SET title = ?, content = ? WHERE id = ?", updatedPost.Title, updatedPost.Content, postId)
+func (r *repositoryImpl) UpdatePost(updatedPost Post, postId int, userId int) (string, error) {
+	result, err := r.db.Exec("UPDATE posts SET title = ?, content = ? WHERE id = ? AND user_id = ?", updatedPost.Title, updatedPost.Content, postId, userId)
 	if err != nil {
 		return "", err
 	}
@@ -80,8 +80,8 @@ func (r *repositoryImpl) UpdatePost(updatedPost Post, postId int) (string, error
 	return "Post updated successfully", nil
 }
 
-func (r *repositoryImpl) DeletePost(postId int) (string, error) {
-	result, err := r.db.Exec("DELETE FROM posts WHERE id = ?", postId)
+func (r *repositoryImpl) DeletePost(postId int, userId int) (string, error) {
+	result, err := r.db.Exec("DELETE FROM posts WHERE id = ? AND user_id = ?", postId, userId)
 	if err != nil {
 		return "", err
 	}
